Fix overzealous heredoc highlighting
As the example Sudoku program demonstrates, the behaviour of heredocs in
GNU APL isn't synonymous with those of other languages. If used inside a
function, the heredoc won't start to collect lines until the function is
called - at which point, it begins collecting whatever's under the call.

Although this is useful, it also makes it impossible for a heredoc to be
used in a function without screwing the document's highlighting. Instead
of removing heredoc formatting altogether, this commit allows authors to
"opt-in" based on whether the terminator contains the language's name:

    text ← ⎕INP "END-OF-HTML"
        <h1> Heading </h1>
        <a href="somewhere.htm"> Link </a>
    END-OF-HTML

Only a handful of languages are recognised in this way: anything else is
regarded as APL code. Authors can enable highlighting for the following:

    ╔══════════════╦═══════════════════════════════════════════════╗
    ║   Language   ║                 Activated by                  ║
    ╟──────────────╫───────────────────────────────────────────────╢
    ║ HTML         ║ HTML, HTM                                     ║
    ║ XML          ║ XML, XSLT, SVG, RSS                           ║
    ║ CSS          ║ CSS, Stylesheet                               ║
    ║ JavaScript   ║ JS, ECMAScript, JavaScript, JScript           ║
    ║ JSON         ║ JSON                                          ║
    ║ Plain text   ║ Raw Text, Plain Text, Raw, Plain              ║
    ╚══════════════╩═══════════════════════════════════════════════╝

The matching is performed case-insensitively, and will also work against
substrings in addition to complete matches (raw text is the exception):

    text ← ⎕INP "HTML"
    text ← ⎕INP "EndHTMLDocument"
    text ← ⎕INP "CSS"

In addition to these changes, a few pattern-matching bugs were fixed:

  * Heredocs needed leading whitespace before they would highlight

  * The closing line was expected to have no other characters except the
    terminating string (whitespace included). GNU APL actually places no
    such restriction, though it discards anything sharing the terminator
    sequence's line.