Rename CSS class applied to ∇ symbols
Since this character is most often used to denote the start and end of a
function's definition, it makes sense to choose a more general name. The
symbol is technically known as a "nabla", often referred to as the "del"
operator in mathematics. I've chosen the former because the latter's too
similar to "delete".