Add impossible end pattern for EOF terminators
Atom allows a rule's "end" pattern to be omitted if the rest of the file
is to be affected. This contradicts the behaviour of other engines, like
TextMate and Lightshow, which require the ending pattern to be included.
To guarantee consistency between rendering engines, an "impossible" rule
has been added to the EOF-terminators introduced in b83e9c8.